backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
scikit-learn
docx2txt
pymupdf
sentence-transformers
openai

python-dotenv
pytest
